os: make 'Genode::Color' C++20 friendly
This patch turns 'Color' from a class to a POD type by replacing
the constructors by the named create functions 'rgb', 'clamped_rgb',
and 'clamped_rgba'. It thereby enables the initialization of Color
values using { .r = ... } syntax and makes the type usable in const
expressions.

It also narrows the type for color components and alpha values to
uint8_t. So possible integer overflows of computed values are detected
by -Wconversion.

As 'Color::rgb(0, 0, 0)' is repeatedly used as a default value, the
patch adds the 'Color::black()' function.

Fixes #5238

+++ repos/gems/src/app/menu_view/label_widget.h
@@ -132,8 +132,8 @@ struct Menu_view::Label_widget : Widget, Cursor::Glyph_position
 		_selections.for_each([&] (Text_selection const &selection) {
 			selection.draw(pixel_surface, alpha_surface, at, text_size.h()); });
 
-		Color const color = _color.color();
-		int   const alpha = color.a;
+		Color   const color = _color.color();
+		uint8_t const alpha = color.a;
 
 		if (alpha) {
 			Text_painter::paint(pixel_surface,
